requirements.txt
anyio==3.7.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.0
exceptiongroup==1.1.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.4.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
